Apply fix for infinite arg with default
Previously when having both infinite and default on 1 arg they would conflict. Namely say I have 2 args, one is a boolean and the other optional infinite roles (empty default). In this situation it is expected that `command on` just sets the boolean arg and keeps roles empty. The bot properly prompts if that is correct giving the user `cancel` or `finish` as options and then the error occurs. Previously when doing `finish` the command would just get canceled due to `user`, but clearly we don't want that since the arg has a default.