Improved typings
This was brougth to my attention when I noticed the recently merged PR #252. While that formed a good start it did optimally use the capabilities of TypeScript. Using the syntax from this PR however that is achieved.

First of all by actually exporting the ArgumentCollectorResult type we
can use it in our own code which in itself is a great addition, even
without 252. And in fact, 252 is pretty useless without it since the
added generic typings weren't being used anywhere. With the addition of
this PR however they are.