fix: convert all arguments to lowercase
If you're checking the lower-case value is included within an array that may not be lower-case, that's pointless and may result in cases where the condition is never true. For example, you want to match an upper-case string in a string[], but because this checks lower-case, it will _never_ be true.